+++ tests/stateful/test_vault.py
@@ -156,7 +156,7 @@ def test_vault(
         st_sender_any = strategy("address")
         st_recip = strategy("address", length=MAX_NUM_SENDERS)
         st_recips = strategy("address[]", length=MAX_NUM_SENDERS, unique=True)
-        st_dstToken = strategy("uint16")
+        st_dstToken = strategy("uint32")
         st_dstAddress = strategy("bytes")
         st_dstChain = strategy("uint32")
         st_message = strategy("bytes")
